fix(run_npu_full): input_ids int32 + A1 verification result
run_npu_full fed input_ids as int64 but the cond axmodel expects int32 (assert fail).
A1 (NPU conditioning) verified on device: full-NPU text->wav works (torch-model-free,
cold ~23s vs slim 37s). A/B vs torch fp32 cond: PERCEPTUALLY EQUAL by ear despite low
full-tensor cos(0.16) — that's masked-padded ±32 saturation the DiT masks out; valid
region cos 0.77 + masking = equivalent audio. NPU cond quantization does NOT break
quality. Separate (orthogonal) issue: both paths have a brief non-JP onset artifact.

+++ e2e_demo/run_npu_full.py
@@ -75,7 +75,7 @@ def main():
     t0 = time.time()
     cond = axe.InferenceSession(args.cond)
     cin = {i.name: i for i in cond.get_inputs()}
-    feed_cond = {"input_ids": ids.astype(np.int64), "mask": text_mask.astype(np.uint8)}
+    feed_cond = {"input_ids": ids.astype(np.int32), "mask": text_mask.astype(np.uint8)}
     feed_cond = {n: feed_cond[n] for n in cin}
     out_names = [o.name for o in cond.get_outputs()]
     cout = cond.run(None, feed_cond)
